fix uncaught promise rejections
Uncaught rejections were from the methods `testManager.runTest` and `testManager.discoverTests`
that are killed when tests complete.
They ought to be handled just to prevent this error, but otherwise there is not any code defect.

I handled it using the same code as in
https://github.com/Microsoft/vscode-python/blob/1cd98b2d3c044a1ed69ca2fe6d548998cc52aabf/src/test/unittests/debugger.test.ts#L150-L156.

I had to disable tslint in a few places because tslint (from the
pre-commit hook) is very strict with promises, see https://stackoverflow.com/q/43980188/5749914.