fix: formatting .pyi files with black
Black has different formatting depending on the file ending.

When formatting a modified buffer or saving a file black will be run
against a temp file that looks roughly like:

```
./.venv/bin/black --diff --quiet ./queryset.pyi.26c0667c29728299036ec32a0f6f7729.tmp
```

The problem with this is that the ending is lost so black will default
to using the `.py` formatting, even when it's a `.pyi` file.

fixes: https://github.com/microsoft/vscode-python/issues/13341