Test envVarsService newline substitution behavior
The behavior of interpreting "\n" inside quoted strings in env files
was neither documented nor tested, so this commit modifies the unit
tests to pin down that behavior.

I also added a test for the example I'm proposing to add to the docs
in PR microsoft/vscode-docs#4888.

I also extended the test for escaping "$" to show that it works even
when the "$" isn't part of something that matches
SUBST_REGEX (i.e. isn't the beginning of something looking like
"${...}").