Disable "snippets" expansion in Jedi LSP
This brings the Jedi based completion experience in line with that
provided by Pylance. Completions now insert only the current symbol
rather than assuming that the user wants to e.g: call that symbol.

This means for example that completing `max` will insert just `max`
rather `max(arg1, arg2)`. While for this case this may be seen as
less useful, it means that insertions in places where a call is
not desired (such as imports and typing contexts) will not be forced
to manually remove the parentheses and template arguments which
might otherwise be inserted.

Users can still use the signature support UI to explore signatures
and of course insertion of an opening parenthesis will still insert
a closing one.

Hopefully this new configuration will be preferable to a majority
of users.

Fixes https://github.com/microsoft/vscode-python/issues/15858