Don't re-run pytest when an exception in a test occurs (#25588)
The general `except` around the pytest run was causing tests to run
twice if the exception handling of pytest is disabled. From the comment
in the code it seems the exception handling is only there for when
reading the test IDs break, so it shouldn't be required around the
pytest main call.

Disabling the exception handling can be practical for debugging tests,
as this starts up the python debugger within vscode. Currently however,
this requires manually patching the run_pytest_script.py, which needs to
be re-done every vscode update.

fixes: https://github.com/microsoft/vscode-python/issues/25656

Co-authored-by: Eleanor Boyd <26030610+eleanorjboyd@users.noreply.github.com>