Allow to use Jedi language server in system environment or virtual environments
As this extension uses an internal Jedi LSP, it is hard for the users to
change the jedi setting to adapt various requirements. And in some cases,
the user require a specific version of the Jedi, which is hard to change
as it is embedded.

Add a new configuration property to allow the users to use Jedi LSP in
their system environment or virtual environments, and auto fallback to
the internal version if the external version is not found.

Signed-off-by: Inochi Amaoto <inochiama@gmail.com>